net: Reorder l2 ovs linux bridge tests
Reorder of `tests/network/l2_bridge/test_l2_ovs_linux_bridge.py`:
- Rename utils.py -> libl2bridge.py
- Move helpers to libl2bridge.py
- Move class VMAttachedToBridge to libl2bridge.py
- Rename test class, edit class description

Add ipv4 mark: tests are adjusted to IPv4.
In order to adjust tests to IPv6, need further
investigtion.
