+++ tests/network/l2_bridge/libl2bridge.py
@@ -2,7 +2,9 @@ import contextlib
 import logging
 import re
 import time
+from ipaddress import ip_interface
 
+from kubernetes.dynamic import DynamicClient
 from ocp_resources.resource import ResourceEditor
 from timeout_sampler import TimeoutExpiredError, TimeoutSampler
 
@@ -22,10 +24,12 @@ from utilities.constants import (
 from utilities.infra import get_pod_by_name_prefix
 from utilities.network import (
     IfaceNotFound,
+    cloud_init_network_data,
     compose_cloud_init_data_dict,
     network_device,
+    ping,
 )
-from utilities.virt import VirtualMachineForTests, fedora_vm_body
+from utilities.virt import VirtualMachineForTests, fedora_vm_body, prepare_cloud_init_user_data
 
 LOGGER = logging.getLogger(__name__)
 
@@ -34,6 +38,7 @@ NETWORK_MANAGER_UNMANAGE_RUNCMD = [
     "sudo systemctl restart NetworkManager",
 ]
 IPV4_ADDRESS_SUBNET_PREFIX_LENGTH = 24
+DHCP_INTERFACE_NAME = "eth3"
 
 
 def _lookup_vmi_interface(vmi, interface_name):
@@ -327,3 +332,166 @@ def create_vm_with_hot_plugged_sriov_interface(
             sriov=True,
         )
         yield vm
+
+
+def wait_for_no_packet_loss_after_connection(src_vm, dst_ip, interface=None):
+    sleep_count_value = 10
+
+    def _get_ping_state():
+        return (
+            ping(
+                src_vm=src_vm,
+                dst_ip=dst_ip,
+                count=sleep_count_value,
+                interface=interface,
+            )
+            == 0
+        )
+
+    try:
+        for sample in TimeoutSampler(
+            wait_timeout=TIMEOUT_2MIN,
+            sleep=sleep_count_value,
+            func=_get_ping_state,
+        ):
+            if sample:
+                return
+    except TimeoutExpiredError:
+        LOGGER.error(f"Ping from {src_vm.name} to {dst_ip} failed.")
+        raise
+
+
+@contextlib.contextmanager
+def bridge_attached_vm(
+    name,
+    namespace,
+    interfaces,
+    ip_addresses,
+    mpls_local_tag,
+    mpls_dest_ip,
+    mpls_dest_tag,
+    mpls_route_next_hop,
+    mpls_local_ip,
+    cloud_init_extra_user_data=None,
+    client=None,
+    node_selector=None,
+):
+    cloud_init_data = _cloud_init_data(
+        vm_name=name,
+        ip_addresses=ip_addresses,
+        mpls_local_ip=mpls_local_ip,
+        mpls_local_tag=mpls_local_tag,
+        mpls_dest_ip=mpls_dest_ip,
+        mpls_dest_tag=mpls_dest_tag,
+        mpls_route_next_hop=mpls_route_next_hop,
+        cloud_init_extra_user_data=cloud_init_extra_user_data,
+    )
+    with VirtualMachineAttachedToBridge(
+        namespace=namespace,
+        name=name,
+        interfaces=interfaces,
+        ip_addresses=ip_addresses,
+        mpls_local_tag=mpls_local_tag,
+        mpls_local_ip=mpls_local_ip,
+        mpls_dest_ip=mpls_dest_ip,
+        mpls_dest_tag=mpls_dest_tag,
+        mpls_route_next_hop=mpls_route_next_hop,
+        client=client,
+        cloud_init_data=cloud_init_data,
+        node_selector=node_selector,
+    ) as vm:
+        yield vm
+
+
+def _cloud_init_data(
+    vm_name,
+    ip_addresses,
+    mpls_local_ip,
+    mpls_local_tag,
+    mpls_dest_ip,
+    mpls_dest_tag,
+    mpls_route_next_hop,
+    cloud_init_extra_user_data,
+):
+    network_data_data = {
+        "ethernets": {
+            "eth1": {"addresses": [f"{ip_addresses[0]}/24"]},
+            "eth2": {"addresses": [f"{ip_addresses[1]}/24"]},
+            "eth4": {"addresses": [f"{ip_addresses[3]}/24"]},
+        },
+    }
+    # Only DHCP server VM (vm-fedora-1) should have IP on eth3 interface
+    if vm_name == "vm-fedora-1":
+        network_data_data["ethernets"][DHCP_INTERFACE_NAME] = {"addresses": [f"{ip_addresses[2]}/24"]}
+
+    # DHCP client VM (vm-fedora-2) should be with dhcp=false, will be activated in test 'test_dhcp_broadcast'.
+    if vm_name == "vm-fedora-2":
+        network_data_data["ethernets"][DHCP_INTERFACE_NAME] = {"dhcp4": False}
+
+    runcmd = [
+        "modprobe mpls_router",  # In order to test mpls we need to load driver
+        "sysctl -w net.mpls.platform_labels=1000",  # Activate mpls labeling feature
+        "sysctl -w net.mpls.conf.eth4.input=1",  # Allow incoming mpls traffic
+        "sysctl -w net.ipv4.conf.all.arp_ignore=1",  # 2 kernel flags are used to disable wrong arp behavior
+        "sysctl -w net.ipv4.conf.all.arp_announce=2",  # Send arp reply only if ip belongs to the interface
+        f"ip addr add {mpls_local_ip} dev lo",
+        f"ip -f mpls route add {mpls_local_tag} dev lo",
+        "nmcli connection up eth4",  # In order to add mpls route we need to make sure that connection is UP
+        f"ip route add {mpls_dest_ip} encap mpls {mpls_dest_tag} via inet {mpls_route_next_hop}",
+        "nmcli connection up eth2",
+        "ip route add 224.0.0.0/4 dev eth2",
+    ]
+
+    cloud_init_data = prepare_cloud_init_user_data(section="runcmd", data=runcmd)
+    cloud_init_data.update(cloud_init_network_data(data=network_data_data))
+
+    if cloud_init_extra_user_data:
+        update_cloud_init_extra_user_data(
+            cloud_init_data=cloud_init_data["userData"],
+            cloud_init_extra_user_data=cloud_init_extra_user_data,
+        )
+
+    return cloud_init_data
+
+
+class VirtualMachineAttachedToBridge(VirtualMachineForTests):
+    def __init__(
+        self,
+        name: str,
+        namespace: str,
+        interfaces: list[str],
+        ip_addresses: list[str],
+        mpls_local_tag: int,
+        mpls_local_ip: str,
+        mpls_dest_ip: str,
+        mpls_dest_tag: int,
+        mpls_route_next_hop: str,
+        client: DynamicClient | None = None,
+        cloud_init_data: dict[str, dict] | None = None,
+        node_selector: dict[str, str] | None = None,
+    ):
+        self.cloud_init_data = cloud_init_data
+        self.mpls_local_tag = mpls_local_tag
+        self.ip_addresses = ip_addresses
+        self.mpls_local_ip = ip_interface(address=mpls_local_ip).ip
+        self.mpls_dest_ip = mpls_dest_ip
+        self.mpls_dest_tag = mpls_dest_tag
+        self.mpls_route_next_hop = mpls_route_next_hop
+
+        networks = {}
+        for network in interfaces:
+            networks.update({network: network})
+
+        super().__init__(  # type: ignore[no-untyped-call]
+            name=name,
+            namespace=namespace,
+            interfaces=interfaces,
+            networks=networks,
+            client=client,
+            cloud_init_data=cloud_init_data,
+            node_selector=node_selector,
+        )
+
+    def to_dict(self) -> None:
+        self.body = fedora_vm_body(name=self.name)
+        super().to_dict()  # type: ignore[no-untyped-call]
